Refactored tests, moved scripts to examples dir.
